requirements.txt
pyyaml
pyzotero
